ecom222/requirements.txt
django==3.0.3
djongo==1.3.1
djangorestframework==3.11.0
Pillow==7.0.0
djangorestframework-simplejwt==4.4.0